+++ appveyor/run_pip.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-import sys
-from pip import main
-
-if __name__ == '__main__':
-    sys.exit(main())